kenni-client-python/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
ecdsa==0.18.0
fastapi==0.100.0
h11==0.14.0
idna==3.4
pyasn1==0.5.0
pydantic==2.0.3
pydantic_core==2.3.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
python-dotenv==1.0.1
